refactor: clean up command parity tests
